+++ src/test/java/org/cojen/dirmi/RestorableTest.java
@@ -571,7 +571,6 @@ public class RestorableTest {
 
     @Test
     public void lenientWithKnownRemoteType() throws Exception {
-        // FIXME: Sometimes fails: java.lang.AssertionError: Didn't reconnect
         lenient(true);
     }